% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babynames}
\alias{babynames}
\title{Ainmneacha Naíonán na hÉireann - Irish Baby Names (1964 - 2018)}
\format{A data frame with four variables: `year`, `sex`, `name` and `n`}
\usage{
babynames
}
\description{
Irish baby name data provided by the Irish CSO dataset VSA05 and VSA12.
This includes all names that occur at least three times in the year.
}
\keyword{datasets}
