% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irishbabynames}
\alias{irishbabynames}
\title{Ainmneacha Naíonán na hÉireann - Irish Baby Names (1964 - 2018)}
\format{A data frame with five variables: `year`, `sex`, `name` and `n`}
\usage{
irishbabynames
}
\description{
Irish baby name data provided by the Irish CSO. This includes all names
that occur at least three times in the year.
}
\keyword{datasets}
