% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irishpopulation}
\alias{irishpopulation}
\title{Irish populationg statistics by gender (1926 - 2018)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6800 rows and 4 columns.}
\usage{
irishpopulation
}
\description{
Statistics collected by the Irish CSO under Annual Population
Estimates PEA11.
}
\keyword{datasets}
