% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migration}
\alias{migration}
\title{Irish migration statistics (1987 - 2018)}
\format{A data frame with five variables: `year`, `sex`, `age_group`, `flow` and `n`}
\usage{
migration
}
\description{
Migration statistics by gender, flow and country of origin/destination.
Statistics collected by the Irish CSO under Annual Population
Estimates PEA03.
}
\keyword{datasets}
