% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Irish populationg statistics by gender (1926 - 2018)}
\format{A data frame with four variables: `year`, `sex`, `age` and `population`}
\usage{
population
}
\description{
Statistics collected by the Irish CSO under Annual Population
Estimates PEA11.
}
\keyword{datasets}
